Fire autocommands before and after writing files in :Acks
I use Syntastic, but don't want it to run when doing multi-file
find/replace via `:Acks` because it adds no value and it slows down the
operation considerably. To give users a good place to hook in to run
their own commands, I am adding FerretWillWrite and FerretDidWrite
`User` autocommands. This will allow me to add some configuration that
disables Syntastic for the operations that I care about.

I decided to use more generic names like `FerretWillWrite` and
`FerretDidWrite` instead of `FerretWillAcks` and `FerretDidAcks` because
I think it more clearly exposes the intent and timing of the
autocommands in the operation.

Fixes #2