Fix handling of multi-word commands
This commit fixes some bugs by reworking the way we use `s:escape`. The
intention of that function was to take a multi-word list of arguments
like:

  -w function!\ s: plugin

And turn it into an escaped string suitable for passing to `system()`.
In this case, `-w` is an option, `function! s:` is the search pattern,
and `plugin` is a directory argument to scope the search to.

We were storing the entire return value of this function in
`g:ferret_lastsearch`, when what we really should have been doing was
storing the search pattern only. As a result, if you tried to use a
multi-word argument list (even something simple like `-w function`) you
would get errors thrown when we tried to eval the saved search
containing a space. Even if errors weren't thrown, the `'hlsearch'`
feature would still be broken for multi-word searches, because it's
unlikely that the buffer contains a string like "-w function").

So, in this commit we rework things in the following ways:

- `s:escape` now does two things: as before, it prepares an escaped
  string ready for use by `system()`, but it now also extracts the
  search pattern and stores it in `g:ferret_lastsearch`
- `s:escape` is renamed to `s:parse` to reflect its new responsibilities
  (figuring out what to do with the arguments list, as opposed to just
  returning a value)
- drop use of `eval()`, which is no longer needed because of the search
  pattern is stored unescaped

Hopefully this isn't too confusing. I'm open to suggestions for a better
name than `s:parse()`; I'd prefer something explicit, but it's hard to
come up with something descriptive that isn't ridiculously verbose (eg
`s:extract_and_store_search_pattern_and_return_escaped_arguments_list`).

CC: @nntn