Show error messages
This is the first step in better error message support.

At the moment we show a "no results" error when there are no results
(fixes https://github.com/wincent/ferret/issues/14), and for other
errors (ie. when every line in the result is "invalid", which basically
means that the command failed), we show the error text (partial fix for
https://github.com/wincent/ferret/issues/10).

Note that actually getting messages to stay on screen is non-trivial in
Vim (see inline comments), so we end up using `input()` to force things.
It seems to be the only way that reliably works across the board (with
and without vim-dispatch).

Still to do: smart heuristics to provide better messages when people
make common mistakes, like misguidedly trying to wrap their search
pattern in quotes.