Special case <> in search queries
I don't fully understand what is going on here, but a search query like
"<Plug>" produces an error:

    ERR: Error: No query. What do you want to search for?

This is exactly the error message that `ag` prints if you supply an
empty search string:

    ag ''

Oddly, searches like "<Plug" and "Plug>" work just fine.

Note that `shellescape` doesn't consider "<" and ">" to be special, even
if you pass in a non-zero `{special}` argument.

Avoid the issue by escaping "<" and ">". All of "<Plug>", "<Plug" and
"Plug>" work then.