Fix for bogus autocomplete issues
This is an attempted fix for an issue I noticed while trying to search
for source code like:

  end # some comment

I can find this on the command-line like this:

  ag '^\s*end\s+#\s+\S+$'

A similar pattern inside Ferret, freaks out because the `-complete=file`
command declaration causes the `#` to get expanded to the alternate file
name, even if you don't actually make use of the completion.

So, provide our own custom completion that only does file globbing for
things that are in file name position.

This mostly seems to work (`:Ack ^\s*end\s*#\s*\S+` and similar variants
work), and the nice thing is I now have a place to add autocomplete for
`ag` options in a later commit.