Fix display of error feedback under vim-dispatch
All messaging and user input has been broken under vim-dispatch since:

https://github.com/tpope/vim-dispatch/commit/c935be9c5e89

That command made our autocmd from run only once instead of twice, but
the time it still ran, it did so under the influence of `:silent`.
Thankfully, Vim provides us with a trump card in the form of
`:unsilent`.