Add g:FerretAutojump
There is a little bit of fragility here because the user can override
the handler behavior with `g:FerretQFHandler` (for example, my use of
`wincmd p` here assumes you've switched focus with `copen` rather than
`cwindow`, but short of removing that functionality I am not sure how
one would avoid this). As it is, the `wincmd` can do the wrong thing, so
this feature is decidedly fragile. eg:

- run a search, and leaving focus in the quickfix listing
- let g:FerretDispatch=0
- let g:FerretAutojump=0
- our wincmd here rotates us out of the quickfix window

Related https://github.com/wincent/ferret/issues/21