Move async functionality into a separate autoloaded file
Some (temporary) duplication here, but the intention is effectively to fork the
plug-in internally. The async part becomes like the "master" branch on which
most work is done and the dispatch/vanilla parts become legacy. I may backport
the occasional fix there, but won't be focussing on them.

This is a first step. More to come.