Allow repeated test runs to work
First run was creating a swapfile that the second run then choked on (it
never got cleaned up because we're killing the host Vim instance via
`tmux kill-session`).

Make the second run ignore the file with `shortmess+=A`. Additionally,
add a `:quitall!` so that the stale swapfile won't be left around in the
first place, at least during successful runs.

Bonus:

- Replace `:set nocompatible` with `-N`; it's not just terser, but it
  helps us avoid ordering issues (because `:set nocompatible` resets a
  bunch of settings, so you have to be careful to call it first).
- `rtp` -> `runtimepath` because it is more transparent to the casual
  reader.