Use ripgrep if available
It's faster.

One weird wrinkle to note: without `--no-heading`, `rg` invoked inside
of Vim via the "async" or "dispatch" methods thinks it is outputting to
a tty (but not via the "vanilla" method).

But note that not all people are going to agree with my precedence order
(`rg` -> `ag` -> `ack` -> `grep`) so I am actually going to make this
overridable in a subsequent commit.