Catch and swallow {c,l}expr errors
This avoids ugly errors when jumping to the first result if the file has
a bad modeline (or something that looks like a modeline, which is how I
ran into this). I think most errors that could happen here are
"innocuous" in nature.

Note that we allow other errors (outside the `{c,l}expr` calls) to
bubble up, as they are likely indicative of real problems.