Avoid 'press ENTER to continue' prompts
I had this hack in place to prevent the error message from getting blown
away before the user sees it, but it is annoying:

    No results for search pattern `foo`: press ENTER to continue

On Vim with 'lambda' and 'timers' support, we can show this instead:

    No results for search pattern `foo`

with no prompt. Strictly speaking, this could have been written without
a lambda expression, but it is tidier this way.