Handle repeated 'g' flags passed to :Acks
This is amazingly edge casey, but the 'g' flag can be repeated, toggling
the effect each time. So:

  :Acks /foo/bar/gggg

Would actually have the effect of turning off global replacement. As
bf9bffbc21ad3 said:

    The docs for `:Acks` promise that:

    > Takes all of the files currently in the |quickfix| listing and
    > performs a substitution of all instances of {pattern} (a standard
    > Vim search |pattern|) by {replacement}.

So in order to be true to the promise of replacing "all instances" we
need to make sure that there is only a single 'g' flag.