Avoid capturing bad 'scrolloff' value
I was finding that my 'scrolloff' was getting permanently set to zero in
the following scenario and perhaps in others:

1. Search for something; the quickfix window gets shown and we save the
   current 'scrolloff' value.
2. Move focus to the quickfix window, at which point we set 'scrolloff'
   to 0.
3. Initiate another search; when the new results show up, we save the
   current value of 'scrolloff' (0), overwriting the real value that we
   originally saved.

The simplest fix, then, seems to be this one: only capture the value
the first time we show the window. This means that if you update the
value and we later do a search, we won't capture the new value, but that
seems less likely to happen than the scenario which we're talking about
in this commit, which happens all the time.