feat: add `<Plug>(FerretBack)`, `<Plug>(FerretBlack)`, `<Plug>(FerretQuack)`
These aren't strictly needed because you can make your own mappings
without them, but the fact that they don't exist may cause confusion,
even for me, as seen in the related issue.

Side note: I'm hacking on a Rust rewrite of `docvim` on this machine
(the Arch Linux box), so I needed to switch back to the `haskell` branch
to be able to generate these updates. These are the commands I ran:

    $ bin/build
    $ bin/docvim -v \
               -c ~/.config/nvim/pack/bundle/opt/ferret \
               ~/.config/nvim/pack/bundle/opt/ferret/README.md \
               ~/.config/nvim/pack/bundle/opt/ferret/doc/ferret.txt

To avoid the need to switch branches in the future, going to do this
ghastly hack now:

    $ cp ./.stack-work/install/x86_64-linux-tinfo6/6496d02f2348a6142a8c6ecb42e6438cffcd197b0d048b4873a88b88b39c503e/7.10.3/bin/docvim ~/bin/

Bit worried that might stop working in the future because it is not
statically linked and I'm not confident that I'll be able to build this
package again in the future:

    $ file ./.stack-work/install/x86_64-linux-tinfo6/6496d02f2348a6142a8c6ecb42e6438cffcd197b0d048b4873a88b88b39c503e/7.10.3/bin/docvim
    ./.stack-work/install/x86_64-linux-tinfo6/6496d02f2348a6142a8c6ecb42e6438cffcd197b0d048b4873a88b88b39c503e/7.10.3/bin/docvim: ELF 64-bit LSB executable, x86-64, version 1 (SYSV), dynamically linked, interpreter /lib64/ld-linux-x86-64.so.2, BuildID[sha1]=82e97c115cb0a6c4fac8eeab9a03a2e0e2325ed9, for GNU/Linux 4.4.0, stripped

There's also:

    $ file ./.stack-work/dist/x86_64-linux-tinfo6/Cabal-1.22.5.0/build/docvim/docvim
    ./.stack-work/dist/x86_64-linux-tinfo6/Cabal-1.22.5.0/build/docvim/docvim: ELF 64-bit LSB executable, x86-64, version 1 (SYSV), dynamically linked, interpreter /lib64/ld-linux-x86-64.so.2, BuildID[sha1]=82e97c115cb0a6c4fac8eeab9a03a2e0e2325ed9, for GNU/Linux 4.4.0, with debug_info, not stripped

(ie. same, with debug symbols)

`ldd` reports these link targets:

    $ ldd ./.stack-work/install/x86_64-linux-tinfo6/6496d02f2348a6142a8c6ecb42e6438cffcd197b0d048b4873a88b88b39c503e/7.10.3/bin/docvim

        linux-vdso.so.1 (0x00007ffe1d7fc000)
        librt.so.1 => /usr/lib/librt.so.1 (0x00007fa5ed433000)
        libutil.so.1 => /usr/lib/libutil.so.1 (0x00007fa5ed42c000)
        libdl.so.2 => /usr/lib/libdl.so.2 (0x00007fa5ed425000)
        libpthread.so.0 => /usr/lib/libpthread.so.0 (0x00007fa5ed404000)
        libgmp.so.10 => /usr/lib/libgmp.so.10 (0x00007fa5ed364000)
        libm.so.6 => /usr/lib/libm.so.6 (0x00007fa5ed220000)
        libc.so.6 => /usr/lib/libc.so.6 (0x00007fa5ed054000)
        /lib64/ld-linux-x86-64.so.2 => /usr/lib64/ld-linux-x86-64.so.2 (0x00007fa5ed460000)

Based on my reading of:

    https://wiki.archlinux.org/title/haskell#Static_linking

it's simply not worth going further down this path. All the proposed
solutions have unpleasant caveats, AFAICT.

Closes: https://github.com/wincent/ferret/issues/79