fix: avoid Vimscript footgun
We can't do a truthiness check for a Vimscript dictionary. ie. if the
quickfix listing has items in it, the `filter()` call would die with:

    E728: Using a Dictionary as a Number

So, we can't just do `v:val`. We can't do `v:val != 0` either, because
Vim will say:

    E735: Can only compare Dictionary with Dictionary

A user-hostile programming language, if ever there was one (well,
AppleScript manages to be even more hostile, but still...)