feat: add `g:FerretCommandNames`
Closes: https://github.com/wincent/ferret/issues/75

Note that as usual, I had headaches getting a working `docvim`
installation in order to update the docs.

In 35757c7c9364d940efd8a46877acde04193819c2, the problem was that the
remote site I was trying to update GHC from was suffering from 503
errors, and I was able to get things working on Arch Linux where I had a
working copy lying around.

Today, back on macOS, I tried a few things:

    cd $PATH_TO_DOCVIM_PROJECT_DIRECTORY
    bin/clean
    rm -r ~/.cabal

but every variation of building I tried would fail with a Happy-related
error. As noted here:

    https://github.com/haskell/happy

> a pre-built Happy is required to build the full version of Happy

In the end, I got it working with:

    cd $PATH_TO_DOCVIM_PROJECT_DIRECTORY
    cabal build -f -bootstrap
    cabal install -f -bootstrap
    bin/docvim \
            -c ~/.config/nvim/pack/bundle/opt/ferret \
            ~/.config/nvim/pack/bundle/opt/ferret/README.md \
            ~/.config/nvim/pack/bundle/opt/ferret/doc/ferret.txt