fix: avoid E474 "Invalid argument" when canceling search job
`l:job` is a string here, but `jobstop()` wants an integer.

When I created the related issue, I forgot that it was already supposed
to do the right thing. I'm guessing that this used to work but at some
point (probably a long time ago), Neovim got stricter about what kind of
argument it would accept for `jobstop()`.

Closes: https://github.com/wincent/ferret/issues/87