feat: complete environment variables and expand them if possible
Existing code completed file names in file name positions on the command
line, and because it used `glob()`, simple cases would be appropriately
handled when invoking the underlying tool. That is, a command-line like:

    :Ack foo $HOME

would end up invoking `rg` (or similar) with the expanded path to
`$HOME` (provided it exists). This happens thanks to the use of
`glob()`, which also handles cases like "{a,b,c*}" (ie. will return "a",
b", and paths that match "c*", if they exist).

At work, I often scope my searches to two specific directories in a
monorepo, and I have an environment variable set like this to make that
easier:

  PROFILING='/path/to/monorepo/some/team/owned/directory /path/to/monorepo/some/other/team/owned/directory'

I have a hack in my dotfiles:

- https://github.com/wincent/wincent/blob/d2efa31568b5b617da8b06b4ae4a4f73054c467e/aspects/nvim/files/.config/nvim/lua/wincent/autocmds.lua#L159-L177

that allows me to define "global command aliases" and type:

    :Ack foo $PROFILING

to find matches in the team directory. How it works: a `CmdlineChanged`
autocommand watches my typing and as soon as it sees `$PROFILING` it
expands the variable. This works, but it looks and feels a bit ugly.

The idea then, is to do this the "right way" inside Ferret, and if I
want to use this approach in other plugins I'll have to extract this
functionality somewhere and use it in multiple places:

- We teach our `:help cmdline-completion` function to complete values
  stating with `$` based on the names of existing environment variables.
- When we actually run the Ferret commands, we soup up our `glob()` call
  to try first a standard glob, then, if that fails and the value we're
  trying to expand starts with a `$`, we try to split it on colons (ie.
  to support variables like `$PATH` and `$MANPATH`), then on commas,
  and finally on whitespace.

Not updating the docs yet because on my work machine where I am writing
this I don't have a working Haskell toolchain, so I can't build the
docs.

The code is ugly and repetitive, but I wanted it to be dumb and simple.

Not sure if this should be gated by an option or not, but I tend to
think the answer is no because the behavior is relatively conservative
(only a small change from what was there before, only applies when
the value starts with `$`, and only actually expands to anything when
the heuristic leads it to find actually-existing paths on the
filesystem).